wallet desktop redesign (#356)
* desktop redesign
