+++ universal-login-wallet/src/styles/components/_balance.sass
@@ -3,6 +3,7 @@
   font-weight: 600
   font-size: 10px
   color: #3D3D41
+  text-transform: uppercase
 
 .balance-box
   padding: 48px 30px 38px
