+++ universal-login-wallet/src/styles/components/_sidebar.sass
@@ -10,7 +10,7 @@
 .sidebar-list-item + .sidebar-list-item
   margin-top: 60px
 
-.sidebar-button
+.sidebar-link
   display: block
   margin: 0 auto
   text-align: center
@@ -20,21 +20,21 @@
 
 .sidebar-button:active
   opacity: 0.6
-  
-.request-funds-button
-  width: 28px
-  height: 23px
-  background: url('../assets/request-funds.svg') center no-repeat
+
+.home-link
+  width: 26px
+  height: 20px
+  background: url('../assets/icons/home.svg') center no-repeat
   background-size: contain
 
-.transfer-funds-button
-  width: 23px
-  height: 24px
-  background: url('../assets/transfer-funds.svg') center no-repeat
-  background-size: contain
-
-.send-invitation-button
-  width: 25px
+.settings-link
+  width: 21px
   height: 21px
-  background: url('../assets/send.svg') center no-repeat
+  background: url('../assets/icons/settings.svg') center no-repeat
+  background-size: contain
+
+.notifications-link
+  width: 23px
+  height: 23px
+  background: url('../assets/icons/bell.svg') center no-repeat
   background-size: contain
