--- universal-login-wallet/src/styles/components/_table.sass
@@ -1,4 +1,0 @@
-.th
-  font-weight: 600
-  font-size: 13px
-  color: #B1B5BF
