+++ universal-login-wallet/src/styles/components/_user-select.sass
@@ -1,6 +1,11 @@
+.user-select-wrapper
+  display: flex
+  justify-content: flex-end
+  padding: 0 90px 0 0
+
 .user-select
   position: relative
-  margin-left: 20px
+  margin-left: auto
 
 .user-select-button
   position: relative
