+++ universal-login-wallet/src/styles/layout/_dashboard.sass
@@ -7,6 +7,7 @@
   width: 100%
 
 .dashboard-content
+  width: 100%
   overflow-y: auto
   padding: 30px 30px 100px
