--- universal-login-wallet/src/styles/layout/_header.sass
@@ -1,18 +1,0 @@
-.header
-  display: flex
-  align-items: center
-  justify-content: flex-end
-  height: 70px
-  width: 100%
-  padding-right: 77px
-  background: #fff
-  z-index: 100
-
-.header-dropdowns
-  position: relative
-  display: flex
-  align-items: center
-  height: 46px
-  padding: 0 14px 0 18px
-  border-right: 1px solid #F2F2FA
-  border-left: 1px solid #F2F2FA
