+++ universal-login-wallet/src/styles/main.sass
@@ -1,7 +1,6 @@
 @import layout/reset
 @import layout/base
 @import layout/dashboard
-@import layout/header
 @import layout/login
 @import layout/transferring
