+++ universal-login-wallet/src/ui/App.tsx
@@ -1,10 +1,11 @@
 import React, {useState} from 'react';
 import {Route, Switch} from 'react-router-dom';
-import Dashboard from './Dashboard/Dashboard';
+import HomeScreen from './Home/HomeScreen';
 import ProgressScreen from './Login/ProgressScreen';
 import NotFound from './NotFound';
 import Login from './Login/Login';
 import {PrivateRoute} from './PrivateRoute';
+import NotificationsScreen from './Notifications/NotificationsScreen';
 
 const App = () => {
   const [authorized, setAuthorized] = useState(false);
@@ -26,7 +27,7 @@ const App = () => {
           path="/"
           render={
             () =>
-            <Dashboard/>
+            <HomeScreen/>
           }
         />
         <PrivateRoute
@@ -35,6 +36,14 @@ const App = () => {
           render={
             () =>
             <ProgressScreen/>
+          }
+        />
+        <PrivateRoute
+          path="/notifications"
+          authorized={authorized}
+          render={
+            () =>
+            <NotificationsScreen />
           }
         />
         <Route component={NotFound}/>
