--- universal-login-wallet/src/ui/Dashboard/Header.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Header = ({children} : {children: any}) => (
-  <div className="header">
-    <div className="header-dropdowns">
-      {children}
-    </div>
-  </div>
-);
-
-export default Header;
